scripts/convert_model/requirements.txt
mlx>=0.22.0
mlx-vlm>=0.0.1
huggingface-hub>=0.25.0
editdistance>=0.8.1
Pillow>=10.0.0
